organize boot/deployment functions
